requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
django-leaflet==0.27.1
django-rest==0.8.6
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
